+++ pom.xml
@@ -30,10 +30,10 @@
     <description>Generate presence metrics for envoys</description>
 
     <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.0.RELEASE</version>
-        <relativePath/> <!-- lookup parent from repository -->
+        <groupId>com.rackspace.salus</groupId>
+        <artifactId>salus-app-base</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
+        <relativePath>../salus-app-base</relativePath>
     </parent>
 
     <properties>
@@ -43,10 +43,6 @@
     </properties>
 
     <dependencies>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-security</artifactId>
-        </dependency>
         <dependency>
             <groupId>org.springframework.retry</groupId>
             <artifactId>spring-retry</artifactId>
@@ -72,16 +68,6 @@
             <optional>true</optional>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.security</groupId>
-            <artifactId>spring-security-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>com.rackspace.salus</groupId>
             <artifactId>salus-common</artifactId>
             <version>0.1.0-SNAPSHOT</version>
@@ -100,6 +86,16 @@
             <groupId>com.github.mxsm</groupId>
             <artifactId>jetcd-launcher</artifactId>
             <version>0.3.6</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.security</groupId>
+            <artifactId>spring-security-test</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
